Updates search to use query parameters for categories
Refactors the search functionality to utilize query parameters for category filtering.
This change improves URL structure and allows for more flexible category management.
The category is now passed as a query parameter instead of part of the path.

Also fixes a bug where the language would be lost on navigation.